Code review, simplified all listeners into 1 listener - EnemyEnterTownListener
